# docs/HRM_swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: HRM_Project
  contact: {}
host: localhost:8081
basePath: /api
securityDefinitions:
  token:
    type: apiKey
    description: ''
    name: token
    in: header
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  '/forms/{id}':
    delete:
      summary: delete form by id
      tags:
        - Form API
      operationId: deleteformbyid
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
    get:
      summary: get form by id
      tags:
        - Form API
      operationId: getformbyid
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
    put:
      summary: approve form of user
      tags:
        - Form API
      operationId: approveformofuser
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/approveformofuserrequest'
      responses:
        '200':
          description: ''
          headers: {}
...
  /forms/all:
    get:
      summary: get all form
      tags:
        - Form API
      operationId: getallform
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: currentPage
          in: query
          required: true
          type: integer
          format: int32
          description: ''
        - name: pageSize
          in: query
          required: true
          type: integer
          format: int32
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
  /accounts:
    post:
      summary: create new account
      tags:
        - Account API
      operationId: createnewaccount
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/createnewaccountrequest'
      responses:
        '200':
          description: ''
          headers: {}
    get:
      summary: get all account
      tags:
        - Account API
      operationId: getallaccount
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: currentPage
          in: query
          required: true
          type: integer
          format: int32
          description: ''
        - name: pageSize
          in: query
          required: true
          type: integer
          format: int32
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
  '/accounts/{id}':
    delete:
      summary: delete account by id
      tags:
        - Account API
      operationId: deleteaccountbyid
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
    put:
      summary: update account by id
      tags:
        - Account API
      operationId: updateaccountbyid
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/updateaccountbyidrequest'
      responses:
        '200':
          description: ''
          headers: {}
    get:
      summary: get account by id
      tags:
        - Account API
      operationId: getaccountbyid
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: ''
      responses:
        '200':
          description: ''
          headers: {}
...
definitions:
...
  approveformofuserrequest:
    title: approveformofuserrequest
    example:
      managerComment: Ok Done
    type: object
    properties:
      managerComment:
        type: string
    required:
      - managerComment
...
  createnewaccountrequest:
    title: createnewaccountrequest
    example:
      userId: cab64ec3-e333-4675-985e-1c7cde4b444b
      username: employee3
      password: director01
    type: object
    properties:
      userId:
        type: string
      username:
        type: string
      password:
        type: string
    required:
      - userId
      - username
      - password
  updateaccountbyidrequest:
    title: updateaccountbyidrequest
    example:
      userId: cab64ec3-e333-4675-985e-1c7cde4b444b
      username: employee3
      password: employee01
    type: object
    properties:
      userId:
        type: string
      username:
        type: string
      password:
        type: string
    required:
      - userId
      - username
      - password
...
security:
  - token: []
...